+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - PASSBOLT-1950: As a user I can see which groups a password is shared with from the sidebar
 - PASSBOLT-1953: As a user I can share a password with a group
+- PASSBOLT-1614: Abstract user/password grid functions into the mad grid library
 
 ## [1.4.0] - 2017-02-07
 ### Fixed